Added Tokenizer Unit Tests
